Renamed EntityId to UniqueId Standarized game data. Removed session and persistent data. Now everything is in the DataProviderLogic and managed internally Added UniqueIdList to have an observable list that is only allowed to be changed on the defined scope (ex: Only Building Logic can change Building data & only Currency Logic can change Currency Data). Added the possibility for the view to observe the data directly on the logic without the possibility to change it

+++ Assets/Src/Configs/BuildingConfigs.cs
@@ -10,11 +10,11 @@ namespace Configs
 	public struct BuildingConfig : IConfig
 	{
 		public GameId Id;
-		public float ProductionAmountBase;
-		public float ProductionAmountIncrease;
+		public int ProductionAmountBase;
+		public int ProductionAmountIncrease;
 		public float ProductionTimeBase;
-		public float UpgradeCostBase;
-		public float UpgradeCostIncrease;
+		public int UpgradeCostBase;
+		public int UpgradeCostIncrease;
 		
 		public int ConfigId => (int) Id;
 	}
